% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs5_data.R
\name{get_acs5_data}
\alias{get_acs5_data}
\title{get_acs5_data}
\usage{
get_acs5_data(period, table, subtable)
}
\arguments{
\item{period}{}

\item{table}{}

\item{subtable}{}
}
\value{

}
\description{
get_acs5_data
}
