% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unemp_from_bls.R
\name{get_unemp_from_bls}
\alias{get_unemp_from_bls}
\title{get unemploymentd data from bls}
\usage{
get_unemp_from_bls(urls)
}
\arguments{
\item{urls}{}
}
\value{

}
\description{
get unemploymentd data from bls
}
